fix: student appt cancel rendering
